requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
arabic-reshaper==2.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asn1crypto==1.5.1
astroid==2.2.5
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5
backcall==0.2.0
backports.zoneinfo==0.2.1
barcode==1.0.2
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==5.0.1
blinker==1.4
boto3==1.20.26
botocore==1.23.26
branca==0.5.0
Brlapi==0.7.0
Brotli==1.0.9
certifi==2022.5.18.1
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
chrome-gnome-shell==0.0.0
cliapp==1.20180812.1
click==8.1.3
cmdtest==0.32+git
colorama==0.4.1
command-not-found==0.3
cryptography==37.0.2
cssselect2==0.6.0
cupshelpers==1.0
dbr==9.2.0.1
dbus-python==1.2.16
debugpy==1.6.2
decorator==5.1.1
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.41.0
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-cleanup==4.0.0
django-cors-headers==3.13.0
django-countries==6.0
django-crispy-forms==1.12.0
django-extensions==3.1.5
django-filter==2.3.0
django-geojson==3.2.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-leaflet==0.28.3
django-location-field==2.1.0
django-on-heroku==1.1.2
django-phonenumber-field==5.2.0
django-rest-framework==0.1.0
django-settings-export==1.2.1
django-sslserver==0.22
django-storages==1.12.3
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.6.0
folium==0.12.1.post1
fonttools==4.33.3
frozenlist==1.3.0
future==0.18.2
geographiclib==1.52
geopy==2.2.0
googlemaps==4.6.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.14.0
idna==2.10
importlib-metadata==2.1.1
importlib-resources==5.8.0
Interface==2.11.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==4.3.21
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
jsonfield==3.1.0
jsonschema==4.6.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.1
lockfile==0.12.2
louis==3.12.0
lxml==4.9.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.12
numpy==1.23.1
oauthlib==3.1.0
olefile==0.46
oscrypto==1.3.0
packaging==20.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pbr==5.8.1
pexpect==4.6.0
phonenumbers==8.12.29
pickleshare==0.7.5
Pillow==8.3.2
pipenv==2022.6.7
platformdirs==2.5.1
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.6.1
psutil==5.9.1
psycopg2==2.9.1
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycodestyle==2.5.0
pycparser==2.21
pycups==1.9.73
pydyf==0.2.0
Pygments==2.12.0
PyGObject==3.36.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
pylint==2.3.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
PyPDF3==1.0.6
pyphen==0.12.0
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-barcode==0.14.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-decouple==3.5
python-dotenv==0.19.2
python-Levenshtein==0.12.2
python3-openid==3.1.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.2.0
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
ratelim==0.1.6
reportlab==3.6.9
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
s3transfer==0.5.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
sip==4.19.21
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
ssh-import-id==5.10
stack-data==0.3.0
stevedore==3.5.0
svglib==1.3.0
systemd-python==234
terminado==0.15.0
tinycss2==1.1.1
tk==0.1.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
ttystatus==0.38
typed-ast==1.4.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritools==4.0.0
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.13.3
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wadllib==1.3.3
wcwidth==0.2.5
weasyprint==55.0
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.6.1
wincertstore==0.2
wrapt==1.11.2
wxPython==4.0.7
xhtml2pdf==0.2.7
xkit==0.0.0
yarl==1.7.2
zipp==3.4.1
zope.event==4.5.0
zope.interface==5.4.0
zope.schema==6.2.0
zopfli==0.2.1
